+++ packages/qfai/src/core/index.ts
@@ -1,4 +1,5 @@
 export * from "./config.js";
+export * from "./decisionGuardrails.js";
 export * from "./ids.js";
 export * from "./report.js";
 export * from "./types.js";